fix script
